backend/requirements.txt
Django>=4.2,<5
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
python-dotenv
django-cors-headers
whitenoise
